Input field for chart speed

+++ Assets/Scripts/Gameplay/Managers/ArcTimingManager.cs
@@ -54,7 +54,7 @@ namespace Arcade.Gameplay
 			{
 				settingVelocity = value;
 				ArcArcManager.Instance.Rebuild();
-				ArcadeComposeManager.Instance.VelocitySlider.SetValueWithoutNotify(value / 30);
+				ArcadeComposeManager.Instance.VelocitySlider.SetValueWithoutNotify(((float)value) / 30);
 			}
 		}
